折る方を決めるカスタムフック
